requirements.txt
asgiref==3.10.0
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
Pillow>=9.0,<12.0
PyJWT==2.10.1
pytz==2025.2
sqlparse==0.5.3
requests
psycopg2-binary==2.9.7